- Rename "export" template overrides to avoid name shadowing by the "partial" schema. - Add `templateOverrides` type to the Options type for generate api in types/index.d.ts - Update README.md file
